pywa,wallpapers,conky,wal-qute,wal-tele,wal-disc,polybar,etc

+++ FilesToCopy/udisks/copy_files.sh
@@ -1,7 +1,5 @@
-#!/bin/env bash
+#!/bin/sh
 
-if [ -f /etc/udisks2/mount_options.conf ]; then
-   sudo rm /etc/udisks2/mount_options.conf 
-fi
+[ -f /etc/udisks2/mount_options.conf ] && sudo rm /etc/udisks2/mount_options.conf 
 
 sudo cp ./* /etc/udisks2/
